Renamed utils.generative into utils.diffusion + fixed broken links
