+++ examples/cfd/flow_reconstruction_diffusion/dataset/dataset.py
@@ -23,7 +23,7 @@ import zipfile
 import numpy as np
 import PIL.Image
 import torch
-from physicsnemo.utils.generative import EasyDict
+from physicsnemo.utils.diffusion import EasyDict
 
 try:
     import pyspng
