+++ examples/cfd/flow_reconstruction_diffusion/training_loop.py
@@ -27,7 +27,7 @@ import psutil
 import torch
 from torch.nn.parallel import DistributedDataParallel
 from training_stats import default_collector, report, report0
-from physicsnemo.utils.generative.utils import (
+from physicsnemo.utils.diffusion.utils import (
     InfiniteSampler,
     check_ddp_consistency,
     construct_class_by_name,
