+++ examples/cfd/flow_reconstruction_diffusion/training_stats.py
@@ -23,7 +23,7 @@ import re
 
 import numpy as np
 import torch
-from physicsnemo.utils.generative.utils import EasyDict, profiled_function
+from physicsnemo.utils.diffusion.utils import EasyDict, profiled_function
 
 # ----------------------------------------------------------------------------
