+++ examples/weather/corrdiff/datasets/hrrrmini.py
@@ -23,7 +23,7 @@ import numpy as np
 from numba import jit, prange
 import xarray as xr
 
-from physicsnemo.utils.generative import convert_datetime_to_cftime
+from physicsnemo.utils.diffusion import convert_datetime_to_cftime
 
 from datasets.base import ChannelMetadata, DownscalingDataset
