+++ examples/weather/corrdiff/helpers/generate_helpers.py
@@ -16,7 +16,7 @@
 
 import datetime
 
-from physicsnemo.utils.generative import convert_datetime_to_cftime
+from physicsnemo.utils.diffusion import convert_datetime_to_cftime
 
 from datasets.dataset import init_dataset_from_config
 from datasets.base import DownscalingDataset
