+++ examples/weather/stormcast/config/sampler/edm_deterministic.yaml
@@ -16,7 +16,7 @@
 
 
 # Sampler args 
-# Below are passed as kwargs to physicsnemo.utils.generative.determinisitic_sampler
+# Below are passed as kwargs to physicsnemo.utils.diffusion.determinisitic_sampler
 # Also supports stochastic sampling via S_churn and related args.
 # See EDM paper for details (https://arxiv.org/abs/2206.00364)
