+++ test/utils/generative/test_stochastic_sampler.py
@@ -47,7 +47,7 @@ class MockNet:
 @import_or_fail("cftime")
 def test_stochastic_sampler(pytestconfig):
 
-    from physicsnemo.utils.generative import stochastic_sampler
+    from physicsnemo.utils.diffusion import stochastic_sampler
 
     net = MockNet()
     latents = torch.randn(2, 3, 448, 448)  # Mock latents
@@ -119,7 +119,7 @@ def test_stochastic_sampler(pytestconfig):
 # The test function for edm_sampler with rectangular domain and patching
 @import_or_fail("cftime")
 def test_stochastic_sampler_rectangle_patching(pytestconfig):
-    from physicsnemo.utils.generative import stochastic_sampler
+    from physicsnemo.utils.diffusion import stochastic_sampler
     from physicsnemo.utils.patching import GridPatching2D
 
     net = MockNet()
